CmakeLists.txt: Only use -mfp16-format when needed
This flag isn't available on any aarch64 compilers, mainly because it's
enabled by default on both clang and later versions of gcc. So exclude
the flag when the compiler doesn't recongnize this, and settle with just
trying to compile some code using __fp16 to make sure that the project
doesn't build unless __fp16 really does work.