Don't redefine PAGE_SIZE and friends
For whatever reason this is already provided as-is on bionic, causing a
redfinition error. I would have just wrapped an ifndef around it, but it
turns out there's a header file we can actually check for and use so
that PAGE_SIZE is always guaranteed to be correct for the system we're
compiling for. And I do some nutty things to define PAGE_SHIFT and take
advantage of constant folding so that we can define it as a constant at
compile-time if we're using bionic's sys/user.h, which has everything
but PAGE_SHIFT.